+++ GrowthBook.Tests/CustomTests/StickyBucketRegressionTests.cs
@@ -3,6 +3,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using FluentAssertions;
 using GrowthBook.MultiUser;
+using GrowthBook.MultiUser.Configuration;
 using GrowthBook.Services;
 using Microsoft.Extensions.Logging.Abstractions;
 using Newtonsoft.Json.Linq;