+++ GrowthBook/Providers/ExperimentEvaluationProvider.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using GrowthBook.Extensions;
-using GrowthBook.MultiUser;
+using GrowthBook.MultiUser.Configuration;
 using GrowthBook.Utilities;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json.Linq;